Fixed some containsIgnoreCase crap
So since possible matches were taken out of the "standbies" pool when a potential match was found, if another potential match was there while the previous one was still being checked, the new potential match would be ignored (since there was nothing matching it in the standbies pool). The old potential match may have been found to not be a match, thus, the actual match would be ignored.

Also there was a character out of bounds error thing. yay.